SQL: add boolean flags for skipping, change params of ticket_state_next()
Add boolean skippable_on_dependent to tbl_ticket_state

Add boolean skip_on_dependent to tbl_project_ticket_state

Change signature of ticket_state_next. Now it takes only the ticket_id as parameter and returns the next state of that ticket based on the current state. Optionally it takes a ticket_state as second parameter and returns the next state based on the assumption the ticket was in that state.